Make util phexdigit array reusable
class template base_uint had its own private lookup table.
This is saving 256 bytes per instantiation.

The result is not spectacular as Beancash-qt has only shrinked of
about 1Kb but it is still valid improvement.

A for loop is replaced with a memset() call.

CBigNum::SetHex() uses the new HexDigit() function.